Add test to verify valid state and disposition combinations
