+++ integration_tests/cypress/fixtures/respondents_sample.csv
@@ -18,183 +18,3 @@
 1000017
 1000018
 1000019
-1000020
-1000021
-1000022
-1000023
-1000024
-1000025
-1000026
-1000027
-1000028
-1000029
-1000030
-1000031
-1000032
-1000033
-1000034
-1000035
-1000036
-1000037
-1000038
-1000039
-1000040
-1000041
-1000042
-1000043
-1000044
-1000045
-1000046
-1000047
-1000048
-1000049
-1000050
-1000051
-1000052
-1000053
-1000054
-1000055
-1000056
-1000057
-1000058
-1000059
-1000060
-1000061
-1000062
-1000063
-1000064
-1000065
-1000066
-1000067
-1000068
-1000069
-1000070
-1000071
-1000072
-1000073
-1000074
-1000075
-1000076
-1000077
-1000078
-1000079
-1000080
-1000081
-1000082
-1000083
-1000084
-1000085
-1000086
-1000087
-1000088
-1000089
-1000090
-1000091
-1000092
-1000093
-1000094
-1000095
-1000096
-1000097
-1000098
-1000099
-1000100
-1000101
-1000102
-1000103
-1000104
-1000105
-1000106
-1000107
-1000108
-1000109
-1000110
-1000111
-1000112
-1000113
-1000114
-1000115
-1000116
-1000117
-1000118
-1000119
-1000120
-1000121
-1000122
-1000123
-1000124
-1000125
-1000126
-1000127
-1000128
-1000129
-1000130
-1000131
-1000132
-1000133
-1000134
-1000135
-1000136
-1000137
-1000138
-1000139
-1000140
-1000141
-1000142
-1000143
-1000144
-1000145
-1000146
-1000147
-1000148
-1000149
-1000150
-1000151
-1000152
-1000153
-1000154
-1000155
-1000156
-1000157
-1000158
-1000159
-1000160
-1000161
-1000162
-1000163
-1000164
-1000165
-1000166
-1000167
-1000168
-1000169
-1000170
-1000171
-1000172
-1000173
-1000174
-1000175
-1000176
-1000177
-1000178
-1000179
-1000180
-1000181
-1000182
-1000183
-1000184
-1000185
-1000186
-1000187
-1000188
-1000189
-1000190
-1000191
-1000192
-1000193
-1000194
-1000195
-1000196
-1000197
-1000198
-1000199
